# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateEnrollmentRequest:
    properties:
      package_id:
        example: 1
        type: integer
    required:
    - package_id
    type: object
...
  dto.EnrollmentAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.StudentEnrollment'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.EnrollmentCheckoutAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.EnrollmentCheckoutData'
      message:
        example: enrollment created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.EnrollmentCheckoutData:
    properties:
      enrollment:
        $ref: '#/definitions/models.StudentEnrollment'
      payment:
        $ref: '#/definitions/models.Payment'
    type: object
  dto.EnrollmentListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.StudentEnrollment'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.ErrorResponse:
    properties:
      errors:
        type: object
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.InvoiceAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.InvoiceData'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.InvoiceData:
    properties:
      created_at:
        type: string
      enrollment:
        $ref: '#/definitions/models.StudentEnrollment'
      id:
        example: 1
        type: integer
      invoice_number:
        example: INV-20260825-0001
        type: string
      issued_at:
        type: string
      payment:
        $ref: '#/definitions/models.Payment'
      payment_id:
        example: 1
        type: integer
      student:
        $ref: '#/definitions/dto.UserResponse'
      updated_at:
        type: string
    type: object
  dto.InvoiceListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.InvoiceData'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.PackageAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.CoursePackage'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.PackageListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.CoursePackage'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.PayRequest:
    properties:
      payment_method:
        allOf:
        - $ref: '#/definitions/models.PaymentMethod'
        enum:
        - BANK_TRANSFER
        - CASH
        example: BANK_TRANSFER
    required:
    - payment_method
    type: object
  dto.PaymentAPIResponse:
    properties:
      data:
        $ref: '#/definitions/models.Payment'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.PaymentCheckoutAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.PaymentCheckoutData'
      message:
        example: payment completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.PaymentListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Payment'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.CoursePackage:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.PackageLevel'
      name:
        type: string
      price:
        type: integer
      status:
        $ref: '#/definitions/models.RecordStatus'
      total_hours:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Payment:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      enrollment_id:
        type: integer
      id:
        type: integer
      paid_at:
        type: string
      payment_code:
        type: string
      payment_method:
        $ref: '#/definitions/models.PaymentMethod'
      status:
        $ref: '#/definitions/models.PaymentStatus'
      updated_at:
        type: string
    type: object
  models.PaymentMethod:
    enum:
    - BANK_TRANSFER
    - CASH
    type: string
    x-enum-varnames:
    - PaymentMethodBankTransfer
    - PaymentMethodCash
...
  models.StudentEnrollment:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      package_id:
        type: integer
      package_name:
        type: string
      package_price:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/models.EnrollmentStatus'
      student_id:
        type: integer
      total_hours:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API Final Project Bootcamp Intensif Golang
  title: Driving Course Management System API
  version: "1.0"
paths:
...
  /api/v1/admin/enrollments:
    get:
      description: List all student enrollments. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EnrollmentListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all student enrollments
      tags:
      - Admin Enrollments
  /api/v1/admin/enrollments/{id}:
    get:
      description: Get student enrollment. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EnrollmentAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get student enrollment
      tags:
      - Admin Enrollments
  /api/v1/admin/invoices:
    get:
      description: List all issued invoices. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all issued invoices
      tags:
      - Admin Invoices
  /api/v1/admin/invoices/{id}:
    get:
      description: Get issued invoice. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get issued invoice
      tags:
      - Admin Invoices
...
  /api/v1/admin/payments:
    get:
      description: List all payments. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all payments
      tags:
      - Admin Payments
  /api/v1/admin/payments/{id}:
    get:
      description: Get payment. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - Admin Payments
...
  /api/v1/student/enrollments:
    get:
      description: List own enrollments. Requires an ACTIVE STUDENT account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EnrollmentListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List own enrollments
      tags:
      - Student Enrollments
    post:
      consumes:
      - application/json
      description: Create enrollment and unpaid payment. Requires an ACTIVE STUDENT
        account.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEnrollmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EnrollmentCheckoutAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create enrollment and unpaid payment
      tags:
      - Student Enrollments
  /api/v1/student/enrollments/{id}:
    get:
      description: Get own enrollment. Requires an ACTIVE STUDENT account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EnrollmentAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own enrollment
      tags:
      - Student Enrollments
  /api/v1/student/enrollments/{id}/payment:
    get:
      description: Get own enrollment payment. Requires an ACTIVE STUDENT account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own enrollment payment
      tags:
      - Student Payments
  /api/v1/student/invoices:
    get:
      description: List own paid invoices. Requires an ACTIVE STUDENT account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List own paid invoices
      tags:
      - Student Invoices
  /api/v1/student/invoices/{id}:
    get:
      description: Get own paid invoice. Requires an ACTIVE STUDENT account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own paid invoice
      tags:
      - Student Invoices
  /api/v1/student/packages:
    get:
      description: List active course packages. Requires an ACTIVE STUDENT account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PackageListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List active course packages
      tags:
      - Student Packages
  /api/v1/student/packages/{id}:
    get:
      description: Get active course package. Requires an ACTIVE STUDENT account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PackageAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active course package
      tags:
      - Student Packages
  /api/v1/student/payments/{id}/pay:
    post:
      consumes:
      - application/json
      description: Simulate payment and issue invoice. Requires an ACTIVE STUDENT
        account.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentCheckoutAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Simulate payment and issue invoice
      tags:
      - Student Payments
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter the token with the Bearer prefix, for example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"